Switch to src layout, move tests to root
