+++ tests/fake_pkg/__init__.pyi
