+++ tests/test_lazy_loader.py
@@ -128,7 +128,7 @@ def test_lazy_attach_returns_copies():
 
 
 def test_attach_same_module_and_attr_name():
-    from lazy_loader.tests import fake_pkg
+    from tests import fake_pkg
 
     # Grab attribute twice, to ensure that importing it does not
     # override function by module
@@ -136,7 +136,7 @@ def test_attach_same_module_and_attr_name():
     assert isinstance(fake_pkg.some_func, types.FunctionType)
 
     # Ensure imports from submodule still work
-    from lazy_loader.tests.fake_pkg.some_func import some_func
+    from tests.fake_pkg.some_func import some_func
 
     assert isinstance(some_func, types.FunctionType)
 
@@ -157,7 +157,7 @@ def test_stub_loading(tmp_path):
 
 
 def test_stub_loading_parity():
-    from lazy_loader.tests import fake_pkg
+    from tests import fake_pkg
 
     from_stub = lazy.attach_stub(fake_pkg.__name__, fake_pkg.__file__)
     stub_getter, stub_dir, stub_all = from_stub
